create shared formatter + refactor
